requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.2
plotly==5.13.0
scipy==1.10.1
pillow